requirements.txt
# Core Flask dependencies
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# OpenAI and Azure dependencies
openai>=1.0.0
tiktoken>=0.5.0
azure-identity>=1.15.0

# AutoGen dependencies (exact versions to match local environment)
autogen-core==0.6.1
autogen-ext==0.6.1
autogen-agentchat==0.6.1

# Data processing and ML dependencies
pydantic>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# LangChain and vector database
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PyTorch and transformers
torch>=2.0.0
transformers>=4.30.0

# Production server
gunicorn>=21.0.0 
firebase_admin
